optimize: toc
By replacing getBoundingClientRect to offsetTop, the function execution time has been reduced signifficantly. The removal of legacy states are moved to the end so no forced reflow is created which improves the performance as well.

There is an unsolved bug that it cannot handle the last heading activation and needs more debugging.